renamed and moved some moduls. added solution for correct import moduls. modiefied test_main
