+++ .gitignore
@@ -158,3 +158,6 @@ cython_debug/
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 .idea/
+
+/home/ksu/PycharmProjects/3_course_Coursework/tests/my_little_tests.py
+
