+++ tests/test_main.py
@@ -1,12 +1,17 @@
-import scr.main as m
+import programs.main as main
 import pytest
 
 
 def test_main_correct():
-    assert m.main() is None
+    assert main.main() is None
 
 
 def test_main_incorrect_path():
-    m.PATH = "operations.json"
+    main.PATH = "operations.json"
     with pytest.raises(FileNotFoundError):
-        m.main()
+        main.main()
+
+
+def test_main_break():
+    main.PATH = "sources/reduced_for_test.json"
+    assert main.main() is None
